+++ paradox-language-support/src/main/gen/com/windea/plugin/idea/paradox/script/psi/ParadoxScriptBlock.java
@@ -25,4 +25,10 @@ public interface ParadoxScriptBlock extends ParadoxScriptValue, PsiListLikeEleme
   @NotNull
   List<PsiElement> getComponents();
 
+  @Nullable
+  ParadoxScriptProperty findProperty(@NotNull String propertyName);
+
+  @Nullable
+  ParadoxScriptValue findValue(@NotNull String value);
+
 }