+++ paradox-language-support/src/main/gen/com/windea/plugin/idea/paradox/script/psi/impl/ParadoxScriptBlockImpl.java
@@ -65,4 +65,16 @@ public class ParadoxScriptBlockImpl extends ParadoxScriptValueImpl implements Pa
     return ParadoxScriptPsiImplUtil.getComponents(this);
   }
 
+  @Override
+  @Nullable
+  public ParadoxScriptProperty findProperty(@NotNull String propertyName) {
+    return ParadoxScriptPsiImplUtil.findProperty(this, propertyName);
+  }
+
+  @Override
+  @Nullable
+  public ParadoxScriptValue findValue(@NotNull String value) {
+    return ParadoxScriptPsiImplUtil.findValue(this, value);
+  }
+
 }