+++ paradox-language-support/src/main/gen/com/windea/plugin/idea/paradox/script/psi/impl/ParadoxScriptPropertyImpl.java
@@ -88,4 +88,16 @@ public class ParadoxScriptPropertyImpl extends ParadoxScriptStubElementImpl<Para
     return ParadoxScriptPsiImplUtil.getTruncatedValue(this);
   }
 
+  @Override
+  @Nullable
+  public ParadoxScriptProperty findProperty(@NotNull String propertyName) {
+    return ParadoxScriptPsiImplUtil.findProperty(this, propertyName);
+  }
+
+  @Override
+  @Nullable
+  public ParadoxScriptValue findValue(@NotNull String value) {
+    return ParadoxScriptPsiImplUtil.findValue(this, value);
+  }
+
 }