+++ paradox-language-support/src/main/kotlin/com/windea/plugin/idea/paradox/script/editor/ParadoxStringScriptPropertyGutterIconRenderer.kt
@@ -16,13 +16,13 @@ class ParadoxStringScriptPropertyGutterIconRenderer(
 	private val properties:Array<ParadoxScriptProperty>
 ): GutterIconRenderer(), DumbAware {
 	companion object{
-		private val _title = message("paradox.script.gutterIcon.property.title")
-		private fun _tooltip(name:String) = message("paradox.script.gutterIcon.property.tooltip", name)
+		private val _title = message("paradox.script.gutterIcon.definition.title")
+		private fun _tooltip(name:String) = message("paradox.script.gutterIcon.definition.tooltip", name)
 	}
 	
 	private val tooltip = _tooltip(name.escapeXml())
 	
-	override fun getIcon() = scriptPropertyGutterIcon
+	override fun getIcon() = stringScriptPropertyGutterIcon
 	override fun getTooltipText() = tooltip
 	override fun getClickAction() = NavigateAction(properties)
 	override fun isNavigateAction() = true